Rename `SharedValue`
